+++ src/externalsystems/OnboardingServiceProvider.Library/OnboardingServiceProviderService.cs
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2025 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -26,6 +26,8 @@ namespace Org.Eclipse.TractusX.Portal.Backend.OnboardingServiceProvider.Library;
 
 public class OnboardingServiceProviderService : IOnboardingServiceProviderService
 {
+    private const string GrantType = "client_credentials";
+    private const string Scope = "openid";
     private readonly ITokenService _tokenService;
 
     /// <summary>
@@ -43,7 +45,9 @@ public class OnboardingServiceProviderService : IOnboardingServiceProviderServic
         {
             TokenAddress = ospDetails.AuthUrl,
             ClientId = ospDetails.ClientId,
-            ClientSecret = ospDetails.ClientSecret
+            ClientSecret = ospDetails.ClientSecret,
+            GrantType = GrantType,
+            Scope = Scope
         };
         using var httpClient = await _tokenService.GetAuthorizedClient<OnboardingServiceProviderService>(settings, cancellationToken)
             .ConfigureAwait(ConfigureAwaitOptions.None);